Добавит файл createPhotos
